+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Show the photo path in the full screen viewer (folder widgets only)
 
+### Fixed
+
+* Fix an issue that could cause interval widgets to become stuck on one photo
+
 ## [v1.42.2] - 2026-05-07
 [v1.42.2]: https://github.com/fibelatti/photo-widget/releases/tag/v1.42.2